% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/settings.R
\name{[.BZAuswertungSettings}
\alias{[.BZAuswertungSettings}
\alias{print.BZAuswertungSettings}
\title{subset method for BZAuswertungSettings class}
\usage{
\method{[}{BZAuswertungSettings}(x, i, ...)

\method{print}{BZAuswertungSettings}(x, ...)
}
\description{
subset method for BZAuswertungSettings class

Print method for BZAuswertungSettings class
}
\keyword{internal}

