% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/settings.R
\name{settings}
\alias{settings}
\title{global settings for BZAuswertung}
\usage{
settings(...)
}
\arguments{
\item{...}{Use parameter value pairs (\code{par1=val1, par2=val2}) to
change a parameter. Use parameter names to request
parameter's value (\code{"par1", "par2"}).}
}
\description{
global settings for BZAuswertung
}
\note{
Currently the following parameters can be changed, ordered by topic.
The default value is shown in the brackets at the end of a line.

 \itemize{
   \item{\code{Bolus.Duration}} {Wie lange wirkt das Bolus Insulin? (\code{4h}) }
   \item{\code{Basal.Duration}} {Wie lange wirkt das Basal Insulin? (\code{12h}) }
   \item{\code{MySugr.ReadHeader}} {Hat die MySugr-Datei eine Kopfzeile? (\code{TRUE}) }
   \item{\code{xxx}} {xyx? (\code{yxy}) }
 }
}
\examples{
\dontrun{
# get current settings
settings()

# get some parameters
settings("Bolus.Duration", "Basal.Duration")

}
}

