% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/settings.R
\name{settingsLoad}
\alias{settingsLoad}
\title{Load BZAuswertung settings}
\usage{
settingsLoad(file)
}
\arguments{
\item{file}{Path of the file to be loaded. If a path is not supplied
an interactive file chooser dialog is opened.}
}
\description{
BZAuswertung settings saved in an a settings file with
the extension \code{.orgset} can be loaded to restore the
settings.
}

