% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/settings.R
\name{settingsSave}
\alias{settingsSave}
\title{Save BZAuswertung settings}
\usage{
settingsSave(file)
}
\arguments{
\item{file}{Path of the file to be saved to. If a path is not supplied
an interactive file saver dialog is opened.}
}
\description{
The current settings of BZAuswertung can be saved into a file with
the extension \code{.orgset}.
}

